file

+++ src/Planet/Bumi.java
@@ -2,7 +2,7 @@ package Planet;
 
 public class Bumi extends Planet{
     
-    public Bumi(int jari,double orbit,double kalarotasi,double kalarevolusi,String namaPlanet) {
+    public Bumi(double jari,double orbit,double kalarotasi,double kalarevolusi,String namaPlanet) {
         super(jari, orbit, kalarotasi, kalarevolusi,namaPlanet);
     }
